fix error with ref
